Basic session authentication
A basic method of session authentication is implemented
and framework for extending it.

Sessions are defined in terms of SessionStorers and SessionAuthenticators

A basic form sessionAuthenticator is implemented as well as a basic
memory sessionStorer and a sample application is provided to demonstrate
how to use them.

However, there should probably be, at a minimum, a database session storer
and for this to be useful.

Fixes #5